+++ src/main/java/com/knowget/knowgetbackend/domain/user/dto/WrittenSuccessCaseDTO.java
@@ -15,7 +15,7 @@ public class WrittenSuccessCaseDTO {
 
 	private String title;
 	private String content;
-	private Short isApproved;
+	private Integer isApproved;
 	private LocalDateTime createdDate;
 
 	public WrittenSuccessCaseDTO(SuccessCase successCase) {
